Update fixedUpdate to use double for deltaTime
Changed the fixedUpdate method to use `double` instead of `float` for deltaTime, improving precision in time updates. Updated relevant signatures and calls in `sdl_game.hpp` and `sdl_game.cpp`. Some external dependencies were also modified.

+++ include/psyengine/sdl_game.hpp
@@ -34,7 +34,7 @@ namespace psyengine
 
     private:
         void handleEvents();
-        static void fixedUpdate(float deltaTime);
+        static void fixedUpdate(double deltaTime);
         static void update(double deltaTime);
         void render(float alpha) const;
